requirements.txt
docker-compose==1.21.2